+++ internal/services/product_service/internal/products/features/deleting_product/v1/delete_product_endpoint.go
@@ -20,7 +20,7 @@ type DeleteProductRequest struct {
 // Validator
 func (e DeleteProductRequest) Schema() v.Schema {
 	return v.Schema{
-		v.F("id", e.Id): v.Gte(int64(0)).Msg("Invalid product id"),
+		v.F("id", e.Id): v.Gte(0).Msg("Invalid product id"),
 	}
 }
